Create React page that gets data from Node server and display videos.
package.json - Add proxy value so requests can be proxied by Webpack
index.html - Add Twitter's widget code so Embeds can be properly displayed
App.js - Created 2 methods, getAllUrls and getEmbed. GetAllUrls gets all the minified video URLs from the server then one by one calls getEmbed on them, which calls to the server to get the actual Embed code. It is then displayed in an embeddedTweets array of divs. Note that for the embed to work the React method 'dangerouslySetInnerHtml' had to be used, or else the code would just be a string.